+++ src/client/java/doctor_m/DOCTORMClient.java
@@ -5,7 +5,5 @@ import net.fabricmc.api.ClientModInitializer;
 public class DOCTORMClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
-
-        System.out.println("Doctor M 客户端已初始化！");
     }
 }