+++ CHANGELOG.md
@@ -8,12 +8,14 @@
 -  50型内饰 / Type 50 interior
 -  克拉拉的塔迪斯内饰 / The Clara TARDIS interior
 -  DOCTOR M创造模式物品栏 / DOCTOR M Creative Mode Inventory
+-  103型塔迪斯与变体NPC / Type 103 Tardis and Variant NPC
 
 ### 修改 / Edit
 -  更新gradle版本为9.2.1-bin，同时加上了MOD依赖项 / Updated the Gradle version to 9.2.1-bin and added the MOD dependencies.
 -  更新去物质枪贴图和模型 / Update DE MAT GUN textures and model
 -  更新拉斯隆之钥的提示 / Prompt to update Rassilon Key
 -  根成就 / Root achievement
+-  主类代码超级优化 / Main class code super optimization
 
 ### 修复 / Repair
 -  进入特兰泽洛可以正确地给予玩家“时候到了？”成就 / Entering Trenzalore can correctly award the player the 'The Time Has Come?' achievement.